Feat: game view에 Timer bar 추가
