+++ Morize/Shared/MorizeApp.swift
@@ -23,9 +23,8 @@ struct MorizeApp: App {
 //            LoginView()
 //                .environmentObject(self.userAuth)
 //                .environmentObject(self.kakaoAuth)
-            MiniGame2B()
+            MiniGame2B(currentPage: .constant(Pages.GamePage),savePhotos:.constant(false))
             
-//            test(currentPage: .constant(Pages.GamePage),savePhotos:.constant(false))
         }
     }
 }
